[review] Fix Approve Article for Submission

+++ src/Services/Review/Review.Application/Dtos/AssetDto.cs
@@ -5,7 +5,7 @@ namespace Review.Application.Dtos;
 public record AssetDto(
 		int Id, 
 		string Name, 
-		byte Number, 
+		int Number, 
 		AssetState State,
 		AssetType Type,
 		FileDto File);
