+++ src/Services/Review/Review.Persistence/Migrations/ReviewDbContextModelSnapshot.cs
@@ -442,8 +442,8 @@ namespace Review.Persistence.Migrations
                         {
                             b1.IsRequired();
 
-                            b1.Property<byte>("Value")
-                                .HasColumnType("tinyint")
+                            b1.Property<int>("Value")
+                                .HasColumnType("int")
                                 .HasColumnName("Number");
                         });
