+++ src/Services/Submission/Submission.Domain/Behaviours/Asset.cs
@@ -10,7 +10,7 @@ public partial class Asset
 				=> $"Articles/{ArticleId}/{Name}/{fileName}";
 
 		//talk - use internal factory method so that the Asset can be created only in the Domain
-		internal static Asset Create(Article article, AssetTypeDefinition type, byte assetCount, IArticleAction action)
+		internal static Asset Create(Article article, AssetTypeDefinition type, int assetCount, IArticleAction action)
 		{
 				//talk - value objects for AssetName & AssetNumber, encapsulate validation						
 				return new Asset()
