+++ src/Services/Submission/Submission.Domain/ValueObjects/AssetName.cs
@@ -7,7 +7,7 @@ public class AssetName: StringValueObject
 		[JsonConstructor]
 		private AssetName(string value) => Value = value;
 
-		public static AssetName Create(AssetTypeDefinition assetType, byte assetCount)
+		public static AssetName Create(AssetTypeDefinition assetType, int assetCount)
 		{
 				if (assetType.AllowsMultipleAssets)
 						return new AssetName($"{assetType.Name.ToString()}_{assetCount + 1}");
