+++ src/Services/Submission/Submission.Persistence/Migrations/SubmissionDbContextModelSnapshot.cs
@@ -303,8 +303,8 @@ namespace Submission.Persistence.Migrations
                         {
                             b1.IsRequired();
 
-                            b1.Property<byte>("Value")
-                                .HasColumnType("tinyint")
+                            b1.Property<int>("Value")
+                                .HasColumnType("int")
                                 .HasColumnName("Number");
                         });
